exposes the deepl error to the developer.
I was getting errors when translating to english. just got the "something went wrong..." message.
What actually happened was that i provided the formality option to an unsupported language (EN in this case).
I did not intially check that this will not work and the error didnt give me any hints.